fix build error by renaming Room namespace to RoomRepository
