+++ el-proyecte-grande-backend/Program.cs
@@ -2,8 +2,8 @@ using el_proyecte_grande_backend.Configurations;
 using el_proyecte_grande_backend.Data;
 using el_proyecte_grande_backend.Repositories.GuestModule;
 using el_proyecte_grande_backend.Repositories.HotelNs;
-using el_proyecte_grande_backend.Repositories.Room;
 using el_proyecte_grande_backend.Repositories.Reservations;
+using el_proyecte_grande_backend.Repositories.RoomRepository;
 using Microsoft.EntityFrameworkCore;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -11,8 +11,6 @@ var builder = WebApplication.CreateBuilder(args);
 // Add services to the container.
 string connectionString = builder.Configuration["ConnectionStrings:GrandeHotelConnection"];
 
-builder.Services.AddScoped<IReservationRepository, ReservationRepository>();
-builder.Services.AddScoped<IRoomRepository, RoomRepository>();
 
 builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
@@ -20,10 +18,13 @@ builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 builder.Services.AddDbContext<GrandeHotelContext>(options =>
     options.UseNpgsql(connectionString));
-builder.Services.AddSingleton<IGuestRepository>(x => 
-    new GuestRepository(x.CreateScope().ServiceProvider.GetRequiredService<GrandeHotelContext>()));
 
+builder.Services.AddSingleton<IGuestRepository>(x =>
+    new GuestRepository(x.CreateScope().ServiceProvider.GetRequiredService<GrandeHotelContext>()));
 builder.Services.AddScoped<IHotelRepository, HotelRepository>();
+builder.Services.AddScoped<IReservationRepository, ReservationRepository>();
+builder.Services.AddScoped<IRoomRepository, RoomRepository>();
+
 builder.Services.AddAutoMapper(typeof(AutoMapperConfiguration));
 
 var app = builder.Build();
