+++ el-proyecte-grande-backend/el-proyecte-grande-backend.csproj
@@ -33,7 +33,6 @@
 
   <ItemGroup>
     <Folder Include="Models\Dtos\" />
-    <Folder Include="Repositories\" />
   </ItemGroup>
 
 </Project>
